Setting a default environment and error feedback
The domain environment now defaults to 'prefix' to stop processing
errors if the $domains['environment'] was not defined. It also ensures
the domain environment variable is lowercase to fix inconsistencies
between 'prefix' and 'Prefix'. Finally, an error message has been
provided for the situation where either prefix or suffix has not been
used.